*Increased compile warning output with -Wextra and fixed all remaining compile warnings *Fixed several bugs i found thanks to the output *Fixed filename bug in explorer (causing crash sometimes) *Added 255 character limit to filenames in the explorer (should be limited to that by the file systems anyway, but to be on the safe side) *Updated Japanese and German language file (thanks hosigumayuugi and ZEN.13)

+++ trunk/source/FTPOperations/ftpii/net.c
@@ -111,7 +111,7 @@ static s32 transfer_exact(s32 s, char *buf, s32 length, transferrer_type transfe
     set_blocking(s, true);
     while (remaining) {
         try_again_with_smaller_buffer:
-        bytes_transferred = transferrer(s, buf, MIN(remaining, NET_BUFFER_SIZE));
+        bytes_transferred = transferrer(s, buf, MIN(remaining, (int) NET_BUFFER_SIZE));
         if (bytes_transferred > 0) {
             remaining -= bytes_transferred;
             buf += bytes_transferred;
